refactor: refactor trip recalculation
