+++ src/routes/trips/+page.svelte
@@ -1,33 +1,14 @@
 <script lang="ts">
-	import type { ChartData, TravelData } from '$tripDomain';
+	import type { TravelData } from '$tripDomain';
 	import { goto } from '$app/navigation';
 	import { TripCard } from '$components';
 	import { allTrips, currentTrip } from '$stores';
+	import { modifyTripData } from '$utils';
 
 	let tripData: Array<TravelData>;
 
 	allTrips.subscribe((storeData) => {
-		tripData = storeData;
-		tripData.map((trip) => {
-			let tripChartData: ChartData = {
-				labels: [],
-				datasets: [
-					{
-						data: [],
-						backgroundColor: []
-					}
-				]
-			};
-			trip.costCategories.map((category) => {
-				tripChartData.labels.push(category.name);
-				tripChartData.datasets[0].data.push(category.totalAmount);
-				tripChartData.datasets[0].backgroundColor.push(category.color);
-				return category;
-			});
-			trip.data = tripChartData;
-			trip.totalCost = tripChartData.datasets[0].data.reduce((partialSum, a) => partialSum + a, 0);
-			return trip;
-		});
+		tripData = modifyTripData(storeData);
 	});
 
 	function onTripCardClick(trip: TravelData) {
