+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export { calculateDate } from './date/calculateDate';
+export { modifyTripData } from './trips/modifytripData';
